Avoid compiler warnings.

+++ libc/stdlib/malloc-standard/mallinfo.c
@@ -88,19 +88,19 @@ void malloc_stats(FILE *file)
     }
 
     mi = mallinfo();
-    fprintf(file, "total bytes allocated             = %10lu\n", (unsigned int)(mi.arena + mi.hblkhd));
-    fprintf(file, "total bytes in use bytes          = %10lu\n", (unsigned int)(mi.uordblks + mi.hblkhd));
-    fprintf(file, "total non-mmapped bytes allocated = %10lu\n", (unsigned int)(mi.arena));
-    fprintf(file, "number of mmapped regions         = %10lu\n", (unsigned int)(mi.hblks));
-    fprintf(file, "total allocated mmap space        = %10lu\n", (unsigned int)(mi.hblkhd));
-    fprintf(file, "total allocated sbrk space        = %10lu\n", (unsigned int)(mi.uordblks));
+    fprintf(file, "total bytes allocated             = %10u\n", (unsigned int)(mi.arena + mi.hblkhd));
+    fprintf(file, "total bytes in use bytes          = %10u\n", (unsigned int)(mi.uordblks + mi.hblkhd));
+    fprintf(file, "total non-mmapped bytes allocated = %10d\n", mi.arena);
+    fprintf(file, "number of mmapped regions         = %10d\n", mi.hblks);
+    fprintf(file, "total allocated mmap space        = %10d\n", mi.hblkhd);
+    fprintf(file, "total allocated sbrk space        = %10d\n", mi.uordblks);
 #if 0
-    fprintf(file, "number of free chunks             = %10lu\n", (unsigned int)(mi.ordblks));
-    fprintf(file, "number of fastbin blocks          = %10lu\n", (unsigned int)(mi.smblks));
-    fprintf(file, "space in freed fastbin blocks     = %10lu\n", (unsigned int)(mi.fsmblks));
+    fprintf(file, "number of free chunks             = %10d\n", mi.ordblks);
+    fprintf(file, "number of fastbin blocks          = %10d\n", mi.smblks);
+    fprintf(file, "space in freed fastbin blocks     = %10d\n", mi.fsmblks);
 #endif
-    fprintf(file, "maximum total allocated space     = %10lu\n", (unsigned int)(mi.usmblks));
-    fprintf(file, "total free space                  = %10lu\n", (unsigned int)(mi.fordblks));
-    fprintf(file, "memory releasable via malloc_trim = %10lu\n", (unsigned int)(mi.keepcost));
+    fprintf(file, "maximum total allocated space     = %10d\n", mi.usmblks);
+    fprintf(file, "total free space                  = %10d\n", mi.fordblks);
+    fprintf(file, "memory releasable via malloc_trim = %10d\n", mi.keepcost);
 }
